refactor: move inflation rewards code out of stake program
